refactor(ds): delete accentColors.ts, add brandColor.ts (Commit 3 of 8)
Drops the per-practice runtime accent-color injection — DS uses a fixed
gold and the visual theme no longer varies per practice. The settings
UI still lets admins set a brand color, and the API still validates the
field, so the pure hex/named validator survives in a new module.

src/shared/utils/brandColor.ts (new):
- Single export: normalizeAccentColor(value?: string | null): string | null
- Accepts named accents ('gold', 'blue', ...) -> their 500-shade hex
- Accepts 3/6-digit hex -> uppercase 6-digit canonical form
- Returns null for invalid input. Callers pair with ?? '#D4AF37'.
- No DOM side effects.

Delete src/shared/utils/accentColors.ts (the runtime palette injector +
applyAccentColor/initializeAccentColor/getCurrentAccentColor/
getTextColorForBackground/getAccentColorDisplayName/ACCENT_COLORS).

9 importers updated:
- src/index.tsx: drop import + the boot-time savedAccent localStorage
  block + the per-org useEffect that called applyAccentColor.
- src/app/MainApp.tsx: drop initializeAccentColor effect; remove the
  now-unused fullAccentColor destructure.
- src/app/WidgetApp.tsx: drop initializeAccentColor effect.
- src/pages/PracticeHomePage.tsx: drop applyAccentColor effect; remove
  the now-unused practiceDetails destructure.
- src/features/settings/pages/PracticePage.tsx: redirect normalizeAccentColor
  import to brandColor; drop the applyAccentColor call + the
  localStorage 'accent-color' writes (dead state).
- src/features/chat/components/WorkspaceSetupSection.tsx: redirect
  normalizeAccentColor import; drop initializeAccentColor effect.
- src/features/chat/pages/hooks/useWorkspaceSetup.ts: redirect import.
- src/features/practice-onboarding/pages/PracticeOnboardingPage.tsx:
  redirect import.
- src/features/practice-onboarding/components/OnboardingDialogs.tsx:
  redirect import.

src/shared/hooks/useWorkspaceRouting.ts: update the dangling comment that
referenced initializeAccentColor.

src/features/media/components/AudioRecordingUI.tsx: the visualizer used
to read --accent-color (a hex string injected by applyAccentColor).
Switch to reading --accent-rgb (the DS RGB triplet defined in
tokens.css) and wrap as rgb(...). Fall back to '#d4af37' (DS gold) if
the var is missing.

tests/component/widget-app.test.tsx: delete the vi.mock for accentColors
and rewrite the accent-theming test as a smoke test for both hex and
named brandColor inputs (since per-practice theming no longer has
observable behavior).

Verified: zero matches for accentColors|applyAccentColor|
initializeAccentColor across src and tests. Build passes. Lint back to
the 1 pre-existing Message.tsx error.

Note: the localStorage key 'accent-color' is no longer written but
older user state still has it. Harmless dead key; will age out.
