+++ src/features/media/components/AudioRecordingUI.tsx
@@ -100,9 +100,10 @@ const AudioRecordingUI: FunctionComponent<AudioRecordingUIProps> = ({
         const barGap = 2;
         const barCount = Math.floor(canvas.width / (barWidth + barGap));
         const radius = 2; // Radius for rounded corners
-        const accentColor = getComputedStyle(document.documentElement).getPropertyValue('--accent-color');
-        
-        ctx.fillStyle = accentColor || '#0ea5e9';
+        const accentRgb = getComputedStyle(document.documentElement).getPropertyValue('--accent-rgb').trim();
+        const accentColor = accentRgb ? `rgb(${accentRgb})` : '#d4af37';
+
+        ctx.fillStyle = accentColor;
         
         function draw() {
             if (!analyserRef.current || !dataArrayRef.current) return;
@@ -188,9 +189,10 @@ const AudioRecordingUI: FunctionComponent<AudioRecordingUIProps> = ({
         const barHeightMultiplier = 30;
         const baseHeight = canvas.height / 4;
         const radius = 2; // Radius for rounded corners
-        const accentColor = getComputedStyle(document.documentElement).getPropertyValue('--accent-color');
-        
-        ctx.fillStyle = accentColor || '#0ea5e9';
+        const accentRgb = getComputedStyle(document.documentElement).getPropertyValue('--accent-rgb').trim();
+        const accentColor = accentRgb ? `rgb(${accentRgb})` : '#d4af37';
+
+        ctx.fillStyle = accentColor;
         
         function animateBarData() {
             const amplitudes = [];
